+++ tests/image_methods_test.py
@@ -206,6 +206,18 @@ def test_chop():
         assert img.height == (h - 10)
 
 
+def test_clahe():
+    if MAGICK_VERSION_NUMBER < 0x700:
+        with raises(WandLibraryVersionError):
+            with Image(filename='rose:') as img:
+                img.clahe(10, 10, 128, 3)
+    else:
+        with Image(filename='rose:') as img:
+            was = img.signature
+            img.clahe(10, 10, 128, 3)
+            assert was != img.signature
+
+
 def test_clut(fx_asset):
     with Image(filename='rose:') as img:
         was = img.signature