+++ docs/changes.rst
@@ -18,6 +18,7 @@ Unreleased.
  - Rewrote :meth:`Image.contrast_stretch() <wand.image.BaseImage.contrast_stretch>`
    method to follow modern CLI behavior.
  - Added :meth:`Image.chop() <wand.image.BaseImage.chop>` method.
+ - Added :meth:`Image.clahe() <wand.image.BaseImage.clahe>` method.
  - Added :meth:`Image.features() <wand.image.BaseImage.features>` method.
  - Added :meth:`Image.forward_fourier_transform() <wand.image.BaseImage.forward_fourier_transform>` method.
  - Added :meth:`Image.inverse_fourier_transform() <wand.image.BaseImage.inverse_fourier_transform>` method.